move google cloud storage integration into its own file
